+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added a field `adjust_instruction_addr: Option<bool>` to `RawFrame` to signal whether the
   frame's instruction address needs to be adjusted for symbolication. ([#948](https://github.com/getsentry/symbolicator/pull/948))
 - Added offline mode and caching to `symbolicli`. ([#967](https://github.com/getsentry/symbolicator/pull/967),[#968](https://github.com/getsentry/symbolicator/pull/968))
+- Support PortablePDB embedded sources. ([#996](https://github.com/getsentry/symbolicator/pull/996))
 - Use `moka` as an in-memory `Cacher` implementation. ([#979](https://github.com/getsentry/symbolicator/pull/979))
 
 ### Internal
@@ -19,9 +20,9 @@
 
 ### Dependencies
 
-- Bump Native SDK from v0.5.0 to v0.5.3 ([#940](https://github.com/getsentry/symbolicator/pull/940))
-  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#053)
-  - [diff](https://github.com/getsentry/sentry-native/compare/0.5.0...0.5.3)
+- Bump Native SDK from v0.5.0 to v0.5.4 ([#940](https://github.com/getsentry/symbolicator/pull/940), [#997](https://github.com/getsentry/symbolicator/pull/997))
+  - [changelog](https://github.com/getsentry/sentry-native/blob/master/CHANGELOG.md#054)
+  - [diff](https://github.com/getsentry/sentry-native/compare/0.5.0...0.5.4)
 
 ## 0.6.0